# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateTaskRequest:
    properties:
      actual_duration:
        maximum: 1440
        minimum: 1
        type: integer
      actual_units:
        minimum: 0
        type: integer
      category:
        type: string
      deadline:
        description: ISO 8601 format
        type: string
      expected_duration:
        maximum: 1440
        minimum: 1
        type: integer
      expected_units:
        maximum: 1000
        minimum: 1
        type: integer
      name:
        type: string
      notes:
        type: string
      start_at:
        maximum: 1439
        minimum: 0
        type: integer
      status:
        enum:
        - CREATED
        - COMPLETED
        - FAILED
        - DEFERRED
        type: string
      task_category:
        enum:
        - ACTION
        - ANCHOR
        - TRANSIT
        type: string
    required:
    - name
    - task_category
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  handlers.UpdateTaskRequest:
    properties:
      actual_duration:
        maximum: 1440
        minimum: 1
        type: integer
      actual_units:
        minimum: 0
        type: integer
      category:
        type: string
      deadline:
        description: ISO 8601 format, empty string to clear
        type: string
      expected_duration:
        maximum: 1440
        minimum: 1
        type: integer
      expected_units:
        maximum: 1000
        minimum: 1
        type: integer
      name:
        type: string
      notes:
        type: string
      start_at:
        maximum: 1439
        minimum: 0
        type: integer
      status:
        enum:
        - CREATED
        - COMPLETED
        - FAILED
        - DEFERRED
        type: string
      task_category:
        enum:
        - ACTION
        - ANCHOR
        - TRANSIT
        type: string
    type: object
  handlers.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Nudge application backend for user management and task tracking.
  title: Nudge API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /tasks:
    get:
      parameters:
      - description: Filter by status
        enum:
        - CREATED
        - COMPLETED
        - FAILED
        - DEFERRED
        in: query
        name: status
        type: string
      - description: Filter by category
        enum:
        - ACTION
        - ANCHOR
        - TRANSIT
        in: query
        name: task_category
        type: string
      - description: Filter by user tag
        in: query
        name: category
        type: string
      - description: Search tasks by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Task creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      parameters:
      - description: Task UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task UUID
        in: path
        name: id
        required: true
        type: string
      - description: Task update payload (all fields optional)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Tasks
securityDefinitions:
  BearerAuth:
    description: 'JWT bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"